+++ frontend/src/scenes/automations/AutomationStepSidebar/AutomationStepMenu.tsx
@@ -3,14 +3,46 @@ import { useActions, useValues } from 'kea'
 import { automationLogic } from '../automationLogic'
 import './AutomationStepConfig.scss'
 import { automationStepConfigLogic, kindToConfig } from './automationStepConfigLogic'
-import { AnyAutomationStep, AutomationStepCategory } from '../schema'
+import { AnyAutomationStep, AutomationStepCategory, AutomationStepKind } from '../schema'
 import { automationStepMenuLogic } from './automationStepMenuLogic'
 import { AutomationStepSidebar } from './AutomationStepSidebar'
+
+const stepOptions: AnyAutomationStep[] = [
+    {
+        kind: AutomationStepKind.EventSource,
+        id: 'new',
+        category: AutomationStepCategory.Source,
+        filters: [],
+    },
+    // { kind: AutomationStepKind.ActionSource, category: AutomationStepCategory.Source },
+    // { kind: AutomationStepKind.PauseForLogic, category: AutomationStepCategory.Logic },
+    // { kind: AutomationStepKind.PauseUntilLogic, category: AutomationStepCategory.Logic },
+    // {
+    //     kind: AutomationStepKind.GithubIssueDestination,
+    //     category: AutomationStepCategory.Destination,
+    // },
+    // {
+    //     kind: AutomationStepKind.UserPropertyDestination,
+    //     category: AutomationStepCategory.Destination,
+    // },
+    // { kind: AutomationStepKind.CohortDestination, category: AutomationStepCategory.Destination },
+    // {
+    //     kind: AutomationStepKind.FeatureFlagDestination,
+    //     category: AutomationStepCategory.Destination,
+    // },
+    { kind: AutomationStepKind.WebhookDestination, id: 'new', category: AutomationStepCategory.Destination },
+    // { kind: AutomationStepKind.SlackDestination, category: AutomationStepCategory.Destination },
+    // { kind: AutomationStepKind.ZapierDestination, category: AutomationStepCategory.Destination },
+    // { kind: AutomationStepKind.EmailDestination, category: AutomationStepCategory.Destination },
+    // {
+    //     kind: AutomationStepKind.InAppMessageDestination,
+    //     category: AutomationStepCategory.Destination,
+    // },
+]
 
 export function AutomationStepMenu(): JSX.Element {
     const { closeMenu } = useActions(automationStepMenuLogic)
     const { setActiveStepId } = useActions(automationStepConfigLogic)
-    const { stepOptions } = useValues(automationStepConfigLogic)
     const { updateStep } = useActions(automationLogic)
 
     return (
